requirements.txt
jmespath==1.0.1
python-dotenv==1.1.1
requests==2.32.3
jsonpath-ng==1.7.0
